Make the test macro include parsing tests
