+++ src/bytecode/parse.rs
@@ -199,103 +199,3 @@ pub fn parse(text: &str) -> Result<Program, ParseError> {
         entry_point: 0,
     })
 }
-
-#[cfg(test)]
-mod test {
-    use super::*;
-
-    #[test]
-    fn test_parse() {
-        use bytecode::Val::*;
-        use bytecode::Instr::*;
-
-        assert_eq!(
-            parse(
-                r#"
-defn f0 2 : 42 69 f1
-x0 := k0
-x1 := k1
-x0 := (x0..x1)
-x1 := k2
-x0 := x1(x0)
-return x0
-
-defn f1 3 : 0 1
-x1 := k0
-x1 := x0[x1]
-x2 := k1
-x2 := x0[x2]
-x0 := x1 + x2
-return x0
-
-defn f2 3 : 1 2 7 15
-x0:= k0
-x1 :=k1
-x0 := x0+ x1
-x0 := x0*x0
-x1 := k2
-x0:=x0 %x1
-x1 := k3
-x0 := x1 / x0
-return x0
-
-defn f3 2 :
-x0 :=read
-x1 := read
-x1 := x1 + x1
-write x1
-write x0
-"#
-            ),
-            Ok(Program {
-                defns: vec![
-                    Defn {
-                        code: vec![
-                            Const(0, 0),
-                            Const(1, 1),
-                            MkTup(0, 0, 1),
-                            Const(1, 2),
-                            Call(0, 1, 0),
-                            Return(Some(0)),
-                        ],
-                        consts: vec![I(42), I(69), C(1)],
-                        local_count: 2,
-                    },
-                    Defn {
-                        code: vec![
-                            Const(1, 0),
-                            IdxTup(1, 0, 1),
-                            Const(2, 1),
-                            IdxTup(2, 0, 2),
-                            Add(0, 1, 2),
-                            Return(Some(0)),
-                        ],
-                        consts: vec![I(0), I(1)],
-                        local_count: 3,
-                    },
-                    Defn {
-                        code: vec![
-                            Const(0, 0),
-                            Const(1, 1),
-                            Add(0, 0, 1),
-                            Mul(0, 0, 0),
-                            Const(1, 2),
-                            Rem(0, 0, 1),
-                            Const(1, 3),
-                            Div(0, 1, 0),
-                            Return(Some(0)),
-                        ],
-                        consts: vec![I(1), I(2), I(7), I(15)],
-                        local_count: 3,
-                    },
-                    Defn {
-                        code: vec![Read(0), Read(1), Add(1, 1, 1), Write(1), Write(0)],
-                        consts: vec![],
-                        local_count: 2,
-                    },
-                ],
-                entry_point: 0,
-            })
-        );
-    }
-}
